pyproject.toml
[project]
name = "spark-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "pyspark>=3.5.5",
    "python-dotenv>=1.1.0",
]
